requirement.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
langdetect==1.0.8
MarkupSafe==1.1.1
nltk==3.5
python-dotenv==0.15.0
regex==2020.10.28
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
tqdm==4.51.0
urllib3==1.25.11
Werkzeug==1.0.1
